+++ .gitignore
@@ -1,4 +1,5 @@
-.vscode
-
-__pycache__
+.vscode/
+__pycache__/
 trivia.code-workspace
+old/
+env/
