+++ app/auth/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+auth_bp = Blueprint('auth', __name__, template_folder='templates')
+from . import routes
+
