+++ app/auth/models.py
@@ -1,41 +1,6 @@
-# importamos la instancia de la BD
-from apptrivia import db
-from werkzeug.security import generate_password_hash, check_password_hash
+from app import db
 from flask_login import UserMixin
-
-class Categoria(db.Model):
-    __tablename__ = 'categoria'
-
-    id = db.Column(db.Integer, primary_key=True)
-    descripcion = db.Column(db.String(64), index=True, unique=True)
-
-    preguntas = db.relationship('Pregunta', backref='categoria', lazy='dynamic')
-
-    def __repr__(self):
-        return f'<Categoria: {self.descripcion}>'
-
-
-class Pregunta(db.Model):
-    __tablename__ = 'pregunta'
-    id = db.Column(db.Integer, primary_key=True)
-    text = db.Column(db.String(255), nullable=False, unique=True)
-
-    categoria_id = db.Column(db.Integer, db.ForeignKey('categoria.id'))
-    respuestas = db.relationship('Respuesta', backref='pregunta', lazy='dynamic')
-
-    def __repr__(self):
-        return f'<Pregunta {self.text}>'
-
-class Respuesta(db.Model):
-    __tablename__ = 'respuesta'
-    id = db.Column(db.Integer, primary_key=True)
-    text = db.Column(db.String(255), nullable=False, unique=True)
-    correcta = db.Column(db.Boolean)
-
-    pregunta_id = db.Column(db.Integer, db.ForeignKey('pregunta.id'))
-
-    def __repr__(self):
-        return f'<Respuesta {self.text}>'
+from werkzeug.security import generate_password_hash, check_password_hash
 
 class Usuario(db.Model, UserMixin):
     __tablename__ = 'usuario'
