+++ app/auth/templates/register.html
