+++ app/errors/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+errors_bp = Blueprint('errors', __name__, template_folder='templates')
+from . import routes
