+++ app/public/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+public_bp = Blueprint('public', __name__, template_folder='templates')
+from . import routes
