+++ app/restricted/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+restricted_bp = Blueprint('restricted', __name__, template_folder='templates')
+from . import routes
