+++ app/restricted/templates/test.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+{% block content %}
+    <h1>Home de la trivia</h1>
+    <h3> TEST: Recurso solo dispoinible para usuarios Admin</h3>
+{% endblock %}
